% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_odin.R
\name{compile_odin}
\alias{compile_odin}
\title{Compile an odin model definition file from inst/odin/}
\usage{
compile_odin(def_file = "discrete_odin_def.R")
}
\arguments{
\item{def_file}{filename of the odin DSL file in inst/odin/}
}
\value{
an odin generator with a `def_filename` attribute
}
\description{
Compile an odin model definition file from inst/odin/
}
