% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_macpan2.R
\name{conv_macpan2}
\alias{conv_macpan2}
\title{Reshape mp_trajectory() output to the long format used by odin/pureR versions}
\usage{
conv_macpan2(traj)
}
\arguments{
\item{traj}{data frame returned by run_simulator_macpan2()}
}
\description{
Reshape mp_trajectory() output to the long format used by odin/pureR versions
}
