% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_odin.R
\name{conv_odin}
\alias{conv_odin}
\title{Convert raw odin output to long or wide format}
\usage{
conv_odin(x, format = c("long", "wide"))
}
\arguments{
\item{x}{matrix returned by run_simulator_odin()}

\item{format}{"long" (default) or "wide"}
}
\description{
Convert raw odin output to long or wide format
}
