% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_macpan2.R
\name{run_simulator_macpan2}
\alias{run_simulator_macpan2}
\title{Run a macpan2 simulator}
\usage{
run_simulator_macpan2(x)
}
\arguments{
\item{x}{macpan2 simulator from make_simulator_macpan2()}
}
\description{
Run a macpan2 simulator
}
