% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_pureR.R
\name{run_simulator_pureR}
\alias{run_simulator_pureR}
\title{Run a pureR simulator}
\usage{
run_simulator_pureR(x)
}
\arguments{
\item{x}{simulator list from make_simulator_pureR()}
}
\description{
Run a pureR simulator
}
