% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_odin.R
\name{stop_both_extinct}
\alias{stop_both_extinct}
\title{Stop when both strains are globally extinct}
\usage{
stop_both_extinct(row)
}
\arguments{
\item{row}{single-row matrix from odin$run(), column names "I[patch,strain]"}
}
\description{
Stop when both strains are globally extinct
}
