% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_run.R
\name{sum_run1}
\alias{sum_run1}
\title{Aggregate a single run to per-step population totals}
\usage{
sum_run1(x, return_type = c("long", "wide"))
}
\arguments{
\item{x}{long-format tibble from discrete_run()}

\item{return_type}{"long" or "wide"}
}
\description{
Aggregate a single run to per-step population totals
}
