% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_run.R
\name{sum_runs}
\alias{sum_runs}
\title{Aggregate a single run to per-step population totals}
\usage{
sum_runs(x)
}
\arguments{
\item{x}{a list of tibbles from a multi-run of discrete_run()}
}
\value{
a single, long tibble with runs indicated by the 'run' variable
}
\description{
Aggregate a single run to per-step population totals
}
